refactor: Separate out test utils from exported ones
