+++ src/a11y/a11y.utils.js
@@ -25,12 +25,3 @@ export function run(domNode, rules = {}) {
     })
   })
 }
-
-export function getAriaLabel(label) {
-  if (!label) return undefined
-
-  if (label && label.length && label[0] === '#') {
-    return { 'aria-labelledby': label.replace(/#/g, '') }
-  }
-  return { 'aria-label': label }
-}
