+++ src/a11y/index.js
@@ -1,1 +1,8 @@
-export { getAriaLabel } from './a11y.utils'
+export function getAriaLabel(label) {
+  if (!label) return undefined
+
+  if (label && label.length && label[0] === '#') {
+    return { 'aria-labelledby': label.replace(/#/g, '') }
+  }
+  return { 'aria-label': label }
+}
